+++ internal/pkg/http/api/session/create/handler.go
@@ -24,25 +24,25 @@ func NewHandler(storage storage.Storage) http.Handler {
 
 func (h *Handler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	if r.Body == nil {
-		errors.NewServerError(uint16(http.StatusBadRequest), "empty request body").RespondWithJSON(w)
+		errors.NewServerError(http.StatusBadRequest, "empty request body").RespondWithJSON(w)
 		return
 	}
 
 	body, readErr := ioutil.ReadAll(r.Body)
 	if readErr != nil {
-		errors.NewServerError(uint16(http.StatusInternalServerError), readErr.Error()).RespondWithJSON(w)
+		errors.NewServerError(http.StatusInternalServerError, readErr.Error()).RespondWithJSON(w)
 		return
 	}
 
 	var req = request{}
 
 	if err := h.json.Unmarshal(body, &req); err != nil {
-		errors.NewServerError(uint16(http.StatusBadRequest), "cannot parse passed json").RespondWithJSON(w)
+		errors.NewServerError(http.StatusBadRequest, "cannot parse passed json").RespondWithJSON(w)
 		return
 	}
 
 	if err := req.validate(); err != nil {
-		errors.NewServerError(uint16(http.StatusBadRequest), "invalid value passed: "+err.Error()).RespondWithJSON(w)
+		errors.NewServerError(http.StatusBadRequest, "invalid value passed: "+err.Error()).RespondWithJSON(w)
 		return
 	}
 
@@ -53,7 +53,7 @@ func (h *Handler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 		time.Second*time.Duration(req.responseDelaySec()),
 	)
 	if sessionErr != nil {
-		errors.NewServerError(uint16(http.StatusInternalServerError), sessionErr.Error()).RespondWithJSON(w)
+		errors.NewServerError(http.StatusInternalServerError, sessionErr.Error()).RespondWithJSON(w)
 		return
 	}