+++ internal/pkg/http/api/session/delete/handler.go
@@ -26,24 +26,24 @@ func NewHandler(storage storage.Storage) http.Handler {
 func (h *Handler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	sessionUUID, sessionFound := mux.Vars(r)["sessionUUID"]
 	if !sessionFound {
-		errors.NewServerError(uint16(http.StatusInternalServerError), "cannot extract session UUID").RespondWithJSON(w)
+		errors.NewServerError(http.StatusInternalServerError, "cannot extract session UUID").RespondWithJSON(w)
 		return
 	}
 
 	// delete session
 	if result, err := h.storage.DeleteSession(sessionUUID); err != nil {
-		errors.NewServerError(uint16(http.StatusInternalServerError), err.Error()).RespondWithJSON(w)
+		errors.NewServerError(http.StatusInternalServerError, err.Error()).RespondWithJSON(w)
 		return
 	} else if !result {
 		errors.NewServerError(
-			uint16(http.StatusNotFound), fmt.Sprintf("session with UUID %s was not found", sessionUUID),
+			http.StatusNotFound, fmt.Sprintf("session with UUID %s was not found", sessionUUID),
 		).RespondWithJSON(w)
 		return
 	}
 
 	// and recorded session requests
 	if _, err := h.storage.DeleteRequests(sessionUUID); err != nil { // TODO delete requests first and ignore error?
-		errors.NewServerError(uint16(http.StatusInternalServerError), err.Error()).RespondWithJSON(w)
+		errors.NewServerError(http.StatusInternalServerError, err.Error()).RespondWithJSON(w)
 		return
 	}