+++ internal/pkg/http/api/session/requests/delete/handler.go
@@ -33,22 +33,22 @@ func NewHandler(storage storage.Storage, broadcaster broadcaster) http.Handler {
 func (h *Handler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	sessionUUID, sessionFound := mux.Vars(r)["sessionUUID"]
 	if !sessionFound {
-		errors.NewServerError(uint16(http.StatusInternalServerError), "cannot extract session UUID").RespondWithJSON(w)
+		errors.NewServerError(http.StatusInternalServerError, "cannot extract session UUID").RespondWithJSON(w)
 		return
 	}
 
 	requestUUID, requestFound := mux.Vars(r)["requestUUID"]
 	if !requestFound {
-		errors.NewServerError(uint16(http.StatusInternalServerError), "cannot extract request UUID").RespondWithJSON(w)
+		errors.NewServerError(http.StatusInternalServerError, "cannot extract request UUID").RespondWithJSON(w)
 		return
 	}
 
 	if deleted, err := h.storage.DeleteRequest(sessionUUID, requestUUID); err != nil {
-		errors.NewServerError(uint16(http.StatusInternalServerError), err.Error()).RespondWithJSON(w)
+		errors.NewServerError(http.StatusInternalServerError, err.Error()).RespondWithJSON(w)
 		return
 	} else if !deleted {
 		errors.NewServerError(
-			uint16(http.StatusNotFound), fmt.Sprintf("request with UUID %s was not found", requestUUID),
+			http.StatusNotFound, fmt.Sprintf("request with UUID %s was not found", requestUUID),
 		).RespondWithJSON(w)
 
 		return