+++ internal/pkg/http/errors/errors_test.go
@@ -11,14 +11,14 @@ func TestNewServerError(t *testing.T) {
 	err := NewServerError(1, "foo")
 
 	assert.False(t, err.Success)
-	assert.Equal(t, uint16(1), err.Code)
+	assert.Equal(t, 1, err.Code)
 	assert.Equal(t, "foo", err.Message)
 }
 
 func TestServerError_JsonCasting(t *testing.T) {
 	t.Parallel()
 
-	err := ServerError{
+	err := serverError{
 		Code:    123,
 		Success: true,
 		Message: "foo",