+++ internal/pkg/http/middlewares/logreq/logreq_test.go
@@ -1,4 +1,4 @@
-package logreq
+package logreq_test
 
 import (
 	"encoding/json"
@@ -9,6 +9,7 @@ import (
 
 	"github.com/kami-zh/go-capturer"
 	"github.com/stretchr/testify/assert"
+	"github.com/tarampampam/webhook-tester/internal/pkg/http/middlewares/logreq"
 	"go.uber.org/zap"
 )
 
@@ -105,7 +106,7 @@ func TestMiddleware(t *testing.T) {
 				log, err := zap.NewProduction()
 				assert.NoError(t, err)
 
-				New(log).Middleware(tt.giveHandler).ServeHTTP(rr, tt.giveRequest())
+				logreq.New(log).Middleware(tt.giveHandler).ServeHTTP(rr, tt.giveRequest())
 			})
 
 			var asJSON map[string]interface{}